+++ CHANGES.md
@@ -23,6 +23,8 @@ To be released.
 
 -  Fixed a bug that `Swarm` reported `TaskCanceledException` as an unknown
    exception while stopping. [[#275]]
+-  Fixed a bug that `Swarm` didn't stop properly during `Swarm.Preload()`.
+   [[#275]]
 
 [#270]: https://github.com/planetarium/libplanet/pull/270
 [#273]: https://github.com/planetarium/libplanet/issues/273